JSON replies with improved handlers
Redid how replies are handled because I didn't realize that http
requests are asyncronous. So now replies are in json and a callback
handler is implemented for any httpConnection function type. It's a bit
complex, but it makes sense in my head, so just ask me if it's a tad
confusing.