Init Register Screen & HttpConnection improvements
Created the new registration screen and implemented the httpConnection
from the Java side.
Also undid a lot of things regarding the httpConnection, most notably,
the removal of the Event interface. After a while, I realized that
creating an interface to be implement for each type of request was
already being taken care of by the pre-existing HttpListener, so I just
moved everything back over to that interface.