Registration and Logins working
Got the login and registration hooks working with the new MessageWindow.
Made several changes to how the MessageWindow works.
A key thing to remember is to always add this window to the stage of
whatever screen you're working with. This enables you to actually see
the screen. Pretty important if you ask me.